cleanup: separate functions
move common functions to its own .h so it's clearer where are they used and don't get copied in all headers
