+++ src/layout/layout.h
@@ -3,6 +3,7 @@
 
 #include "../streamtypes.h"
 #include "../vgmstream.h"
+#include "../util/reader_sf.h"
 
 /* blocked layouts */
 void render_vgmstream_blocked(sample_t* buffer, int32_t sample_count, VGMSTREAM* vgmstream);
