+++ src/meta/nub.c
@@ -1,5 +1,6 @@
 #include "meta.h"
 #include "../coding/coding.h"
+#include "../util/endianness.h"
 
 
 static void load_name(char* name, size_t name_size, STREAMFILE* sf, int big_endian, int total_subsongs, int target_subsong);
@@ -33,7 +34,7 @@ VGMSTREAM* init_vgmstream_nub(STREAMFILE* sf) {
         goto fail;
 
     /* sometimes LE [Soul Calibur: Broken Destiny (PSP), Tales of Vesperia (PS4) */
-    big_endian = guess_endianness32bit(0x18, sf);
+    big_endian = guess_endian32(0x18, sf);
     if (big_endian) {
         read_32bit = read_32bitBE;
     } else{
@@ -249,7 +250,7 @@ VGMSTREAM* init_vgmstream_nub_wav(STREAMFILE* sf) {
     if (read_32bitBE(0x00,sf) != 0x77617600) /* "wav\0" "*/
         goto fail;
 
-    if (guess_endianness32bit(0x1c, sf)) {
+    if (guess_endian32(0x1c, sf)) {
         read_32bit = read_32bitBE;
         read_16bit = read_16bitBE;
     } else {
@@ -313,7 +314,7 @@ VGMSTREAM* init_vgmstream_nub_vag(STREAMFILE* sf) {
     if (read_32bitBE(0x00,sf) != 0x76616700) /* "vag\0" */
         goto fail;
 
-    if (guess_endianness32bit(0x1c, sf)) {
+    if (guess_endian32(0x1c, sf)) {
         read_32bit = read_32bitBE;
     } else {
         read_32bit = read_32bitLE;
@@ -598,7 +599,7 @@ VGMSTREAM* init_vgmstream_nub_is14(STREAMFILE* sf) {
     if (read_32bitBE(0x00,sf) != 0x69733134) /* "is14" */
         goto fail;
 
-    if (guess_endianness32bit(0x1c, sf)) {
+    if (guess_endian32(0x1c, sf)) {
         read_32bit = read_32bitBE;
     } else{
         read_32bit = read_32bitLE;
