+++ src/util/endianness.h
@@ -2,6 +2,7 @@
 #define _UTIL_ENDIAN_H
 
 #include "../streamfile.h"
+#include "reader_get.h"
 
 typedef uint32_t (*read_u32_t)(off_t, STREAMFILE*);
 typedef  int32_t (*read_s32_t)(off_t, STREAMFILE*);
@@ -11,8 +12,17 @@ typedef float (*read_f32_t)(off_t, STREAMFILE*);
 
 typedef  int16_t (*get_s16_t)(const uint8_t*);
 
-//todo move here
-#define guess_endian32 guess_endianness32bit
-#define guess_endian16 guess_endianness16bit
+/* guess byte endianness from a given value, return true if big endian and false if little endian */
+static inline int guess_endian16(off_t offset, STREAMFILE* sf) {
+    uint8_t buf[0x02];
+    if (read_streamfile(buf, offset, 0x02, sf) != 0x02) return -1; /* ? */
+    return get_u16le(buf) > get_u16be(buf) ? 1 : 0;
+}
+
+static inline int guess_endian32(off_t offset, STREAMFILE* sf) {
+    uint8_t buf[0x04];
+    if (read_streamfile(buf, offset, 0x04, sf) != 0x04) return -1; /* ? */
+    return get_u32le(buf) > get_u32be(buf) ? 1 : 0;
+}
 
 #endif
